requirements.txt
numpy>=1.26
matplotlib>=3.8
typer>=0.12
imageio>=2.34
PyYAML>=6.0
rich>=13.7
Pillow>=10.0
opencv-python-headless>=4.10
cvxpy>=1.5
pytest>=7.4
hypothesis>=6.98
black>=24.3
ruff>=0.6
mypy>=1.10
mkdocs==1.6.0
mkdocs-material==9.5.35
mkdocs-autorefs==1.2.0
mkdocs-git-revision-date-localized-plugin==1.2.4
mkdocs-minify-plugin==0.8.0
pymdown-extensions==10.8.1